Refactor test cases to use advanced features
Googletest does natively support running the same test for different
parameter combinations or for different types. This is better than
manually looping over the needed values as googletest will automatically
indicate the currently used parameters/types in the test output, making
it easier to make sense of the output, should a test fail.