CI: Update cmake and use Ninja on all platforms
Older versions of cmake don't properly support the SYSTEM property for
include directories with the VS generator that is used by default on
Windows, which leads to a bunch of unrelated warning messages that cause
the build to fail.

This issue should be fixed with cmake ~3.23, but to be sure we not only
upgrade cmake, but also resort to using Ninja instead.