Rewrote expression formatter
The older formatter had a couple of bugs. In particular, the used
approach was unfit to deal with expressions like (a + b) * c, e.g.
where the lhs of the multiplication needs parenthesis.

Besides fixing this principle issue with the old approach, the new
version now also correctly encodes the order of
multiplication/contraction in the involved sub-expressions, which will
become important once we are dealing with contractions of more than two
tensors and the different factorization possibilities are associated
with different contraction costs.