Ensure substituting doesn't invalidate replaced Node
It turns out that making sure that the TreeNode that represents the root
of the sub-tree that gets substituted with some expression remains valid
in the sense that the ID of that TreeNode remains pointing to an node
that is part of the tree after the substitution.

This is important for allowing substitutions while iterating the tree
(aka: concurrent modification of the tree), provided that the
substitution aims at the current element in the iteration.